fix: don't use ts-node for running tests (#48)
