+++ test/fixtures/bad-tilejson/bad-attribution-not-string.json
