+++ test/fixtures/bad-tilejson/bad-bounds-array-non-number.json
