+++ test/fixtures/bad-tilejson/bad-bounds-not-array.json
