+++ test/fixtures/bad-tilejson/bad-center-array-non-number.json
