+++ test/fixtures/bad-tilejson/bad-center-not-array.json
