+++ test/fixtures/bad-tilejson/bad-data-array-non-string.json
