+++ test/fixtures/bad-tilejson/bad-description-not-string.json
