+++ test/fixtures/bad-tilejson/bad-empty-object.json
