+++ test/fixtures/bad-tilejson/bad-legend-not-string.json
