+++ test/fixtures/bad-tilejson/bad-max-not-int.json
