+++ test/fixtures/bad-tilejson/bad-max-not-number.json
