+++ test/fixtures/bad-tilejson/bad-max-too-high.json
