+++ test/fixtures/bad-tilejson/bad-max-too-low.json
