+++ test/fixtures/bad-tilejson/bad-min-not-int.json
