+++ test/fixtures/bad-tilejson/bad-min-not-number.json
