+++ test/fixtures/bad-tilejson/bad-min-too-high.json
