+++ test/fixtures/bad-tilejson/bad-min-too-low.json
