+++ test/fixtures/bad-tilejson/bad-name-not-string.json
