+++ test/fixtures/bad-tilejson/bad-not-an-object.json
