+++ test/fixtures/bad-tilejson/bad-scheme-not-string.json
