+++ test/fixtures/bad-tilejson/bad-scheme-not-xyz-or-tms.json
