+++ test/fixtures/bad-tilejson/bad-template-not-string.json
