+++ test/fixtures/bad-tilejson/bad-tiles-array-non-string.json
