+++ test/fixtures/bad-tilejson/bad-version-not-string.json
