+++ test/fixtures/bad-tilejson/invalid-url2.json
