+++ test/fixtures/bad-tilejson/invalid-url3.json
