+++ test/fixtures/good-stylejson/good-simple-raster.json
