+++ test/fixtures/good-stylejson/good-simple-vector.json
