+++ test/fixtures/good-tilejson/mapbox_raster_tilejson.json
