+++ test/fixtures/mbtiles/vector/trails-pbf.mbtiles
