+++ test/test-helpers/mock-server.js
@@ -1,15 +1,15 @@
-import { rest } from 'msw'
-import { createHash } from 'crypto'
-import { TileJSON } from '../lib/tilejson'
+const { setupServer } = require('msw/node')
+const { rest } = require('msw')
+const { createHash } = require('crypto')
 
-export const handlers = [
+const handlers = [
   rest.get(
     // based on fixtures/good-tilejson/mapbox_raster_tilejson.json
     'http://*.tiles.mapbox.com/v3/aj.1x1-degrees/:zoom/:x/:y',
     async (req, res, ctx) => {
       const { x, y: tempY, zoom } = req.params
 
-      const y = (tempY as string).split('.')[0]
+      const y = /** @type {string} */ (tempY).split('.')[0]
 
       // First 8 bytes identify a PNG datastream: https://www.w3.org/TR/PNG/#5PNG-file-signature
       const body = Buffer.from([
@@ -43,9 +43,9 @@ export const handlers = [
   rest.get('https://api.mapbox.com/v4/:tileset', async (req, res, ctx) => {
     const { tileset } = req.params
 
-    const mbTilesetId = (tileset as string).replace('.json', '')
+    const mbTilesetId = /** @type {string} */ (tileset).replace('.json', '')
 
-    const tilejson: TileJSON = {
+    const tilejson = {
       id: mbTilesetId,
       tilejson: '2.2.0',
       format: 'pbf',
@@ -82,8 +82,15 @@ export const handlers = [
   }),
 ]
 
-// An adjusted version of https://github.com/jshttp/etag/blob/4664b6e53c85a56521076f9c5004dd9626ae10c8/index.js#L39
-function createETag(entity: string | Buffer): string {
+//
+/**
+ * An adjusted version of
+ * https://github.com/jshttp/etag/blob/4664b6e53c85a56521076f9c5004dd9626ae10c8/index.js#L39
+ *
+ * @param {string | Buffer} entity
+ * @returns {string}
+ */
+function createETag(entity) {
   const hash = createHash('sha1')
     .update(entity.toString('utf8'), 'utf8')
     .digest('base64')
@@ -97,6 +104,13 @@ function createETag(entity: string | Buffer): string {
   return `"${len.toString(16)}-${hash}"`
 }
 
-function convertParamToNumber(param: string | readonly string[]): number {
+/**
+ * @param {string | readonly string[]} param
+ * @returns number
+ */
+function convertParamToNumber(param) {
   return Number.parseInt(Array.isArray(param) ? param[0] : param, 10)
 }
+
+// This configures a request mocking server with the given request handlers.
+module.exports = setupServer(...handlers)
