+++ tsconfig.json
@@ -19,18 +19,9 @@
     "resolveJsonModule": true,
     "pretty": true,
     "outDir": "dist",
-    "typeRoots": [
-      "./node_modules/@types",
-      "./types"
-    ],
+    "typeRoots": ["./node_modules/@types", "./types"],
     "allowJs": true,
+    "sourceMap": true
   },
-  "include": [
-    "src/**/*",
-    "types/**/*",
-  ],
-  "exclude": [
-    "src/**/*.test.ts",
-    "src/mocks/**/*",
-  ]
+  "include": ["src/**/*", "types/**/*"]
 }
